+++ HISTORY.md
@@ -26,10 +26,14 @@
     * significantly sped-up rendering morphs representing data inside table cells
     * support testing selectors (translatable text) for equality with text and numbers without needing to "unselect" them
     * added support for color slot default values
+    * replaced "Object" type input slot with "Color" type in the input slot dialog ("Object" is now in the "special slots" menu)
 * **Notable Fixes:**
     * fixed sorting the distribution of compound non-list data (blocks, colors)
 * **Translation Updates:**
     * German
+
+### 2025-04-16
+* byob: replaced "Object" type input slot with "Color" type in the input slot dialog ("Object" is now in the "special slots" menu)
 
 ### 2025-04-15
 * blocks, objects: added support for color slot default values in primitives